Fix the indexing to be much simpler.
We just grab all the forecast data but for the analysis data we skip
until we hit the 6am data so it matches the forecast data. This should
work with those data which are sampled only 4 times a day (as opposed to
the majority which are sampled 8 times a day).